created component-folder + moved components + renamed main.jsx = index.jsx
