+++ app.html
@@ -11,6 +11,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script src="./src/app/main.jsx"></script>
+    <script src="./src/app/index.jsx"></script>
 </body>
 </html>
