+++ src/app/components/app.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./app.css";
+// import "./app.css";
 import Logo from "./logo";
 import UserForm from "./user-form";
 import NavBar from "./nav-bar";
