+++ src/app/components/logo.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import logo from "./images/recycle.png";
-import "./app.css";
+import logo from "../images/recycle.png";
+// import "./app.css";
 
 export default class Logo extends React.Component {
